src/ml_strategy_manager.py
"""Machine learning-based trading strategy management.

Handles loading and execution of ML models for trading decisions
including Random Forest and LSTM-based strategies.
"""

import logging

import yaml


class MLStrategyManager:
    """Manages machine learning-based trading strategies."""

    def __init__(self, db, config):
        """Initialize MLStrategyManager.

        Args:
            db: Database manager instance
            config: Configuration dictionary
        """
        self.db = db
        self.config = config
        self.logger = logging.getLogger(__name__)
        self.strategies = []
        self.load_strategies()

    def load_strategies(self):
        """Load ML strategies from config"""
        try:
            with open("src/config/config.yaml", "r", encoding="utf-8") as file:
                config = yaml.safe_load(file)
            for strategy_config in config.get("ml", []):
                self.strategies.append(strategy_config)
                self.logger.debug("Loaded ML strategy: %s", strategy_config["name"])
        except (IOError, ValueError) as e:
            self.logger.error("Failed to load ML strategies: %s", e)

    def generate_signals(self, symbol=None):
        """Generate signals for ML strategies (placeholder)"""
        self.logger.debug("Generating ML signals for %s", symbol or "all symbols")
        return []
