requirements.txt
# Core Trading
MetaTrader5>=5.0.0
pandas>=1.3.0

# Technical Analysis & Indicators
ta>=0.10.0
scikit-learn>=1.0.0
numba>=0.56.0

# Backtesting Metrics
empyrical>=0.5.5
backtesting>=0.3.3

# Visualization & Charts
plotly>=5.0.0
kaleido>=0.2.1
matplotlib>=3.4.0
Pillow>=9.0.0

# UI Framework - Web
Flask>=2.0.0
Flask-CORS>=3.0.0

# UI Framework - Desktop (optional, kept for backward compatibility)
PyQt5>=5.15.0
PyQtWebEngine>=5.15.0

# Configuration & Data
PyYAML>=5.4.0

# Testing
pytest>=9.0.0
pytest-cov>=4.0.0

# Utilities
numpy>=1.20.0,<2.0.0  # NumPy < 2.0 for empyrical compatibility
requests>=2.27.0
python-dateutil>=2.8.2

